Extract query attribute grammar
